Move norm_id into representation
